Condensing files into one component
